A4x12_TablaTemperaturasFor_Modif

+++ programacion/eclipse-workspace/UD01/src/bucles/actividades/A4x12_TablaTemperaturasFor_Modif.java
@@ -10,9 +10,9 @@ public class A4x12_TablaTemperaturasFor_Modif {
 		for (int celsius = -20; celsius <= 100; celsius += 10) {
 			final double K = 273.15;
 			int fahrenheit = (celsius * 9) / 5 + 32;
-			double kelvin = (celsius + K);
+			int kelvin = (int) (celsius + K);
 			
-			System.out.printf("%7d %10d %6.0f\n", celsius, fahrenheit, kelvin);
+			System.out.printf("%7d %10d %6d\n", celsius, fahrenheit, kelvin);
 
 		}
 	}
